move(cosmo): move tests to private API
