+++ astropy/cosmology/_src/tests/conftest.py
@@ -2,5 +2,5 @@
 
 """Configure the tests for :mod:`astropy.cosmology`."""
 
-from astropy.cosmology.tests.helper import clean_registry  # noqa: F401
+from astropy.cosmology._src.tests.helper import clean_registry  # noqa: F401
 from astropy.tests.helper import pickle_protocol  # noqa: F401
