+++ astropy/cosmology/flrw/tests/test_base.py
@@ -17,15 +17,15 @@ import pytest
 import astropy.constants as const
 import astropy.units as u
 from astropy.cosmology import FLRW, FlatLambdaCDM, LambdaCDM, Planck18
-from astropy.cosmology.core import _COSMOLOGY_CLASSES, dataclass_decorator
-from astropy.cosmology.flrw.base import _a_B_c2, quad
-from astropy.cosmology.tests.helper import get_redshift_methods
-from astropy.cosmology.tests.test_core import (
+from astropy.cosmology._src.tests.helper import get_redshift_methods
+from astropy.cosmology._src.tests.test_core import (
     CosmologyTest,
     FlatCosmologyMixinTest,
     invalid_zs,
     valid_zs,
 )
+from astropy.cosmology.core import _COSMOLOGY_CLASSES, dataclass_decorator
+from astropy.cosmology.flrw.base import _a_B_c2, quad
 from astropy.tests.helper import assert_quantity_allclose
 from astropy.utils.compat.optional_deps import HAS_PANDAS, HAS_SCIPY
