+++ astropy/cosmology/flrw/tests/test_w0cdm.py
@@ -7,8 +7,8 @@ import pytest
 
 import astropy.units as u
 from astropy.cosmology import FlatwCDM, wCDM
+from astropy.cosmology._src.tests.test_core import ParameterTestMixin, valid_zs
 from astropy.cosmology.parameter import Parameter
-from astropy.cosmology.tests.test_core import ParameterTestMixin, valid_zs
 from astropy.tests.helper import assert_quantity_allclose
 from astropy.utils.compat.optional_deps import HAS_SCIPY
