+++ astropy/cosmology/flrw/tests/test_w0wacdm.py
@@ -7,8 +7,8 @@ import pytest
 
 import astropy.units as u
 from astropy.cosmology import Flatw0waCDM, Planck18, w0waCDM
+from astropy.cosmology._src.tests.test_core import ParameterTestMixin
 from astropy.cosmology.parameter import Parameter
-from astropy.cosmology.tests.test_core import ParameterTestMixin
 from astropy.tests.helper import assert_quantity_allclose
 from astropy.utils.compat.optional_deps import HAS_SCIPY
