+++ astropy/cosmology/flrw/tests/test_wpwazpcdm.py
@@ -8,8 +8,8 @@ import pytest
 import astropy.cosmology.units as cu
 import astropy.units as u
 from astropy.cosmology import FlatwpwaCDM, wpwaCDM
+from astropy.cosmology._src.tests.test_core import ParameterTestMixin
 from astropy.cosmology.parameter import Parameter
-from astropy.cosmology.tests.test_core import ParameterTestMixin
 from astropy.tests.helper import assert_quantity_allclose
 from astropy.utils.compat.optional_deps import HAS_SCIPY
